Hedge against content that is not a string-part 2
Same as before, but for the convert25 method.

In the unlikely event that a user has a puppet define with a parameter
named 'content' and this is false, puppet catalog diff will crash and
burn in a fiery blaze reminiscent of the Hindenburg disaster.

Less dramatically, this unhelpful error is what you get:

err: can't convert false into String
err: Try 'puppet help catalog diff' for usage