Add option --show_resource_diff
Currently, it is difficult to read the current resource
difference output format (esp when there are loads of
params for the resources).

This commit adds an option: --show_resource_diff which indicates
that resource differences should be displayed as a diff between
the resources and not as each entire resource.