Protect against diff edge case
this patch adds a discovered edge case.

It is possible for resource parameters to be
different, but for diff not to detec the
differences.

An example is the case of a resource parameter
value of 'true' vs. true. This patch adds
code to protect against this edge case (previously,
It was failing with 'undefined method `join' for nil:NilClass'.

This commit handles thid edge case, and prints inspect of the
parameter hashes.