Add option --exclude_classes
Parameterized class declarations show
up as resources in the diff output.

It makes sense to ignore classes (esp when using
the catalog diff tool to validate a refactor) b/c
they don't effect the end state of a system (they
merely contain resources, which is what you really
should care about when doing a diff)

This commit adds an option which strips all classes
out of the list of resources being compared.