Intial refactor for YAML output
Prior to this commit this face used `puts` for output. This is an
intial commit for the refactor, to convert this face to work with
`when_rendering` and thus the `--render-as yaml` for programatic
comparision and formatting of the output. This commit does not
include the resource difference migration yet as that is going to
take a little work to move the strings to ruby objects