Scape node name from new filename
Prior to this commit the key at the top of the hash was the filename
this commit changes it to the "new" filename minus the ext. This
is assuming that something akin to `facter fqdn` was used in the
creation of the node files. Thus this fqdn or certname is the key