Add Seed support to face
Prior to this commit one had to run puppet from the commandline
in a loop. This commit adds a seed option which allows catalogs
to be compiled filtered by facts. This allows for the first
half of the process to be automated on the servers and filters
be put on the test group of diff nodes