Fix ordering issue
Prior to this commit we were handing things in backwards
This commit updates the code to have the same convention as the
files when handling directories. i.e. old new. This needs more
cleanup as the whole to/from thing is being used inchangablity with
from and to but does not need to be