Add Support for Environments to seed
Prior to this commit we always compiled the catalog using the pro
duction environment. This commit adds support to ask the yaml termi
nus for the last environment that the node checked in with. We
then use that environment to compile the catalog

This whole set of code needs to be moved into its own class so
that we can look at multithreading.