Validate node is active
Prior to this commit we would compile the catalog by
searching only the yaml terminus. The issue with this
was that if the node had checked into the master but
had then later been deactived, when we processed this
list we would export resources for this node as if
it was active. Futher we would duplicate said external
nodes if the yaml terminus resturned a node it
did not have a yaml file. That last bit is highly
unlikely but happened during the intial `puppet master --compile`
style of the code