Add md5 hash of errors minus node_name
Prior to this commit we did not match errors that were close
to each other. This commit md5s the error minus the node_name
assume that that is the only unique content. This should
increase the depth of the hash in a meaningful way

The example errors should then be more useful then the md5s