Include list of `all_changed_nodes` in json report (#34)
Currently, `puppet-catalog-diff-viewer` only looks at the nodes in
`most_differences` to decide which nodes to display as 'with changes'
and 'no changes'.  By default, this list is limited to just 10 nodes
with the most differences.  Including a complete list of changed nodes
in the json (regardless of `--changed_depth`) will make fixing
puppet-catalog-diff-viewer much easier.